fix: entry_key not have '.' returns 500

+++ auth/v2_del_key.py
@@ -9,23 +9,26 @@ api_key_header = APIKeyHeader(name=API_KEY_NAME, auto_error=True)
 
 def Delete_API_key_AuthV2(entry_key: str, req_permission: str = "a", security: str = Security(api_key_header)):
     try:
-        uid_part, key_part = entry_key.split('.', 1)
-        conn = sqlite3.connect('/app/data/api_keys.db')
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM api_keys WHERE uid = ?", (uid_part,))
-        if cursor.rowcount == 0:
-            conn.close()
-            raise HTTPException(status_code=404, detail="API Key not found")
+        if "." in entry_key:
+            uid_part, key_part = entry_key.split('.', 1)
+            conn = sqlite3.connect('/app/data/api_keys.db')
+            cursor = conn.cursor()
+            cursor.execute("DELETE FROM api_keys WHERE uid = ?", (uid_part,))
+            if cursor.rowcount == 0:
+                conn.close()
+                raise HTTPException(status_code=404, detail="API Key not found")
+            else:
+                conn.commit()
+                conn.close()
+                del entry_key
+                del key_part
+                del uid_part
+                return {
+                    "status": "200",
+                    "message": "API Key deleted successfully"
+                }
         else:
-            conn.commit()
-            conn.close()
-            del entry_key
-            del key_part
-            del uid_part
-            return {
-                "status": "200",
-                "message": "API Key deleted successfully"
-            }
+            raise HTTPException(status_code=404, detail="API Key not found")
     except HTTPException as http_exc:
         raise http_exc
     except Exception as e:
